refactor open_dataset
