+++ tests/test_xarray.py
@@ -33,39 +33,36 @@ def test_open_dataset_root() -> None:
 
 
 def test_open_dataset_orbit() -> None:
-    annotation_path = (
+    manifest_path = (
         DATA_FOLDER
         / "S1B_IW_SLC__1SDV_20210401T052622_20210401T052650_026269_032297_EFA4.SAFE"
-        / "annotation"
-        / "s1b-iw1-slc-vv-20210401t052624-20210401t052649-026269-032297-004.xml"
+        / "manifest.safe"
     )
-    res = xr.open_dataset(annotation_path, engine="sentinel-1", group="orbit")  # type: ignore
+    res = xr.open_dataset(manifest_path, engine="sentinel-1", group="IW2/orbit")  # type: ignore
 
     assert isinstance(res, xr.Dataset)
     assert set(res.dims) == {"time"}
 
 
 def test_open_dataset_attitude() -> None:
-    annotation_path = (
+    manifest_path = (
         DATA_FOLDER
         / "S1B_IW_SLC__1SDV_20210401T052622_20210401T052650_026269_032297_EFA4.SAFE"
-        / "annotation"
-        / "s1b-iw1-slc-vv-20210401t052624-20210401t052649-026269-032297-004.xml"
+        / "manifest.safe"
     )
-    res = xr.open_dataset(annotation_path, engine="sentinel-1", group="attitude")  # type: ignore
+    res = xr.open_dataset(manifest_path, engine="sentinel-1", group="IW2/attitude")  # type: ignore
 
     assert isinstance(res, xr.Dataset)
     assert set(res.dims) == {"time"}
 
 
 def test_open_dataset_gcp() -> None:
-    annotation_path = (
+    manifest_path = (
         DATA_FOLDER
         / "S1B_IW_SLC__1SDV_20210401T052622_20210401T052650_026269_032297_EFA4.SAFE"
-        / "annotation"
-        / "s1b-iw1-slc-vv-20210401t052624-20210401t052649-026269-032297-004.xml"
+        / "manifest.safe"
     )
-    res = xr.open_dataset(annotation_path, engine="sentinel-1", group="gcp")  # type: ignore
+    res = xr.open_dataset(manifest_path, engine="sentinel-1", group="IW2/gcp")  # type: ignore
 
     assert isinstance(res, xr.Dataset)
     assert set(res.dims) == {"azimuth_time", "slant_range_time"}
